feat(server): Add route to get single patent - PST-155
